requirements.txt
pytesseract>=0.3.10
transformers>=4.38.0
torch>=2.1.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.66.0
jiwer>=3.0.0
